PR comments based refactoring
