--- Example/Byndyusoft.ApiClient.Example.Benchmark/Byndyusoft.ApiClient.Example.Benchmark.csproj
@@ -1,23 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <OutputType>Exe</OutputType>
-        <RootNamespace>Byndyusoft.ApiClient.Example.Benchmark</RootNamespace>
-    </PropertyGroup>
-    <PropertyGroup>
-        <PlatformTarget>AnyCPU</PlatformTarget>
-        <DebugType>pdbonly</DebugType>
-        <DebugSymbols>true</DebugSymbols>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-        <Optimize>true</Optimize>
-        <Configuration>Release</Configuration>
-        <IsPackable>false</IsPackable>
-    </PropertyGroup>
-    <ItemGroup>
-        <PackageReference Include="BenchmarkDotNet" Version="0.14.0" />
-        <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="8.0.11" />
-    </ItemGroup>
-    <ItemGroup>
-      <ProjectReference Include="..\Byndyusoft.ApiClient.Example.Server\Byndyusoft.ApiClient.Example.Server.csproj" />
-    </ItemGroup>
-</Project>
