--- Example/Byndyusoft.ApiClient.Example.Benchmark/Program.cs
@@ -1,17 +1,0 @@
-namespace Byndyusoft.ApiClient.Example.Benchmark
-{
-    using BenchmarkDotNet.Configs;
-    using BenchmarkDotNet.Running;
-
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var config = DefaultConfig.Instance;
-            var summary = BenchmarkRunner.Run<FormattersBenchmark>(config, args);
-
-            // Use this to select benchmarks from the console:
-            // var summaries = BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args, config);
-        }
-    }
-}
