--- Example/Byndyusoft.ApiClient.Example.Contracts/Byndyusoft.ApiClient.Example.Contracts.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <RootNamespace>Byndyusoft.ApiClient.Example.Contracts</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Byndyusoft.ApiClient.Example.Models\Byndyusoft.ApiClient.Example.Models.csproj" />
-    </ItemGroup>
-
-</Project>
