--- Example/Byndyusoft.ApiClient.Example.Server/Byndyusoft.ApiClient.Example.Server.csproj
@@ -1,31 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-	<PropertyGroup>
-		<OutputType>Exe</OutputType>
-		<TargetFramework>net8.0</TargetFramework>
-		<RootNamespace>Byndyusoft.ApiClient.Example.Server</RootNamespace>
-		<GenerateDocumentationFile>true</GenerateDocumentationFile>
-		<Nullable>enable</Nullable>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Asp.Versioning.Mvc" Version="8.1.0" />
-		<PackageReference Include="Asp.Versioning.Mvc.ApiExplorer" Version="8.1.0" />
-        <PackageReference Include="Byndyusoft.AspNetCore.Instrumentation.Tracing" Version="1.0.0" />
-		<PackageReference Include="Byndyusoft.AspNetCore.Mvc.Formatters.MessagePack" Version="0.4.0" />
-		<PackageReference Include="Byndyusoft.AspNetCore.Mvc.Formatters.ProtoBuf" Version="0.2.0" />
-        <PackageReference Include="Byndyusoft.Logging" Version="3.0.1" />
-        <PackageReference Include="Byndyusoft.Logging.OpenTelemetry" Version="3.0.1" />
-        <PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.11.1" />
-        <PackageReference Include="OpenTelemetry.Exporter.Jaeger" Version="1.5.1" />
-        <PackageReference Include="OpenTelemetry.Exporter.Prometheus" Version="1.3.0-rc.2" />
-        <PackageReference Include="OpenTelemetry.Instrumentation.AspNetCore" Version="1.11.0" />
-		<PackageReference Include="protobuf-net.AspNetCore" Version="3.2.46" />
-		<PackageReference Include="Swashbuckle.AspNetCore" Version="7.2.0" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\Byndyusoft.ApiClient.Example.Client\Byndyusoft.ApiClient.Example.Client.csproj" />
-	</ItemGroup>
-
-</Project>
