--- Example/Byndyusoft.ApiClient.Example.Server/Properties/launchSettings.json
@@ -1,13 +1,0 @@
-{
-  "profiles": {
-    "Byndyusoft.ApiClient.Example": {
-      "commandName": "Project",
-      "launchBrowser": true,
-      "launchUrl": "swagger",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      },
-      "applicationUrl": "https://localhost:5001;http://localhost:5000"
-    }
-  }
-}
