+++ example/Api.Client/Byndyusoft.ApiClient.Example.Client.csproj
@@ -1,0 +1,20 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <RootNamespace>Byndyusoft.ApiClient.Example.Client</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\src\Byndyusoft.ApiClient\Byndyusoft.ApiClient.csproj" />
+      <ProjectReference Include="..\Api.Contracts\Byndyusoft.ApiClient.Example.Contracts.csproj" />
+    </ItemGroup>
+</Project>
