+++ example/Api.Contracts/Byndyusoft.ApiClient.Example.Contracts.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <RootNamespace>Byndyusoft.ApiClient.Example.Contracts</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Models\Byndyusoft.ApiClient.Example.Models.csproj" />
+    </ItemGroup>
+
+</Project>
