+++ example/Api.Contracts/PersonModelListRoutes.cs
@@ -1,0 +1,15 @@
+﻿namespace Byndyusoft.ApiClient.Example.Contracts
+{
+    public class PersonModelListRoutes
+    {
+        public const string PersonModelListPrefix = "Person";
+        public const string GetEveryPersonCmd = "GetEveryPerson";
+        public const string AddListCmd = "AddList";
+        public const string DeleteListCmd = "DeleteList";
+        public const string GetPersonCmd = "GetPerson";
+        public const string AddPersonCmd = "AddPerson";
+        public const string ReplacePersonCmd = "ReplacePerson";
+        public const string UpdatePersonCmd = "UpdatePerson";
+        public const string DeletePersonCmd = "DeletePerson";
+    }
+}
