+++ example/Api/Byndyusoft.ApiClient.Example.Server.csproj
@@ -1,0 +1,36 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<TargetFramework>net8.0</TargetFramework>
+		<RootNamespace>Byndyusoft.ApiClient.Example.Server</RootNamespace>
+		<GenerateDocumentationFile>true</GenerateDocumentationFile>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Asp.Versioning.Mvc" Version="8.1.0" />
+		<PackageReference Include="Asp.Versioning.Mvc.ApiExplorer" Version="8.1.0" />
+		<PackageReference Include="Byndyusoft.AspNetCore.Instrumentation.Tracing" Version="1.0.0" />
+		<PackageReference Include="Byndyusoft.AspNetCore.Mvc.Formatters.MessagePack" Version="0.4.0" />
+		<PackageReference Include="Byndyusoft.AspNetCore.Mvc.Formatters.ProtoBuf" Version="0.2.0" />
+		<PackageReference Include="Byndyusoft.Logging" Version="3.0.1" />
+		<PackageReference Include="Byndyusoft.Logging.OpenTelemetry" Version="3.0.1" />
+		<PackageReference Include="Npgsql.OpenTelemetry" Version="9.0.2" />
+		<PackageReference Include="OpenTelemetry" Version="1.10.0" />
+		<PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" Version="1.10.0" />
+		<PackageReference Include="OpenTelemetry.Exporter.Prometheus.AspNetCore" Version="1.8.0-rc.1" />
+		<PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.10.0" />
+		<PackageReference Include="OpenTelemetry.Instrumentation.AspNetCore" Version="1.10.1" />
+		<PackageReference Include="OpenTelemetry.Instrumentation.Http" Version="1.10.0" />
+		<PackageReference Include="OpenTelemetry.Instrumentation.Runtime" Version="1.10.0" />
+		<PackageReference Include="protobuf-net.AspNetCore" Version="3.2.46" />
+		<PackageReference Include="Serilog.Extensions.Hosting" Version="9.0.0" />
+		<PackageReference Include="Swashbuckle.AspNetCore" Version="7.2.0" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\Api.Contracts\Byndyusoft.ApiClient.Example.Contracts.csproj" />
+	</ItemGroup>
+
+</Project>
