+++ example/Api/Infrastructure/OpenTelemetryExtensions/AspNetCoreInstrumentationOptionsExtensions.cs
