+++ example/Api/Infrastructure/Swagger/ApplicationBuilderExtensions.cs
