+++ example/Api/Infrastructure/Swagger/ServiceCollectionExtensions.cs
