+++ example/Api/appsettings.Development.json
@@ -1,0 +1,15 @@
+{
+  "ConnectionStrings": {
+    "Main": "Server=localhost;Port=5432;Database={DB};User Id={UserID};Password={Password};POOLING=True;MINPOOLSIZE=1;MAXPOOLSIZE=1024;"
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "OtlpExporterOptions": {
+    "Endpoint": "http://localhost:4317"
+  }
+}
