+++ example/Benchmark/Program.cs
@@ -1,0 +1,14 @@
+namespace Byndyusoft.ApiClient.Example.Benchmark
+{
+    using BenchmarkDotNet.Configs;
+    using BenchmarkDotNet.Running;
+
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            var config = DefaultConfig.Instance;
+            var summary = BenchmarkRunner.Run<FormattersBenchmark>(config, args);
+        }
+    }
+}
