+++ example/Models/Byndyusoft.ApiClient.Example.Models.csproj
