+++ example/Tests/JsonFormatterTests.cs
@@ -1,0 +1,15 @@
+namespace Byndyusoft.ApiClient.Example.Tests;
+
+using Client;
+using Xunit.Abstractions;
+
+public class JsonFormatterTests : FormattersTests
+{
+    public JsonFormatterTests(ITestOutputHelper testOutputHelper) : base(testOutputHelper)
+    {
+        TestSubject = new PersonApi(
+            Client,
+            _clientSettings
+        );
+    }
+}
