+++ example/Tests/MessagePackFormatterTests.cs
@@ -1,0 +1,18 @@
+namespace Byndyusoft.ApiClient.Example.Tests;
+
+using System.Net.Http.MessagePack;
+using System.Net.Http.MessagePack.Formatting;
+using Client;
+using Xunit.Abstractions;
+
+public class MessagePackFormatterTests : FormattersTests
+{
+    public MessagePackFormatterTests(ITestOutputHelper testOutputHelper) : base(testOutputHelper)
+    {
+        TestSubject = new PersonApi(
+            Client,
+            _clientSettings,
+            new MessagePackMediaTypeFormatter(MessagePackDefaults.SerializerOptions)
+        );
+    }
+}
